Reformat application submission function
